+++ src/icalendar/__init__.py
@@ -73,10 +73,12 @@ from icalendar.prop import (
     vSkip,
     vText,
     vTime,
+    vUid,
     vUnknown,
     vUri,
     vUTCOffset,
     vWeekday,
+    vXmlReference,
 )
 from icalendar.prop.conference import Conference
 from icalendar.prop.image import Image
@@ -155,9 +157,11 @@ __all__ = [
     "vText",
     "vTime",
     "vUTCOffset",
+    "vUid",
     "vUnknown",
     "vUri",
     "vWeekday",
+    "vXmlReference",
     "version",
     "version_tuple",
 ]