+++ src/icalendar/prop/__init__.py
@@ -47,12 +47,13 @@ from __future__ import annotations
 import base64
 import binascii
 import re
+import uuid
 from datetime import date, datetime, time, timedelta, timezone
 from typing import Any, ClassVar, Optional, Tuple, Union
 
 from icalendar.caselessdict import CaselessDict
 from icalendar.enums import Enum
-from icalendar.error import JCalParsingError
+from icalendar.error import InvalidCalendar, JCalParsingError
 from icalendar.parser import Parameters, escape_char, unescape_char
 from icalendar.parser_tools import (
     DEFAULT_ENCODING,
@@ -191,8 +192,6 @@ class vBoolean(int):
     BOOL_MAP = CaselessDict({"true": True, "false": False})
 
     def __new__(cls, *args, params: dict[str, Any] | None = None, **kwargs):
-        if params is None:
-            params = {}
         self = super().__new__(cls, *args, **kwargs)
         self.params = Parameters(params)
         return self
@@ -246,8 +245,6 @@ class vText(str):
         /,
         params: dict[str, Any] | None = None,
     ):
-        if params is None:
-            params = {}
         value = to_unicode(value, encoding=encoding)
         self = super().__new__(cls, value)
         self.encoding = encoding
@@ -265,7 +262,12 @@ class vText(str):
         ical_unesc = unescape_char(ical)
         return cls(ical_unesc)
 
-    from icalendar.param import ALTREP, LANGUAGE, RELTYPE
+    @property
+    def ical_value(self) -> str:
+        """The string value of the text."""
+        return str(self)
+
+    from icalendar.param import ALTREP, GAP, LANGUAGE, RELTYPE, VALUE
 
     def to_jcal(self, name: str) -> list:
         """The jCal representation of this property according to :rfc:`7265`."""
@@ -277,8 +279,6 @@ class vText(str):
     def examples(cls):
         """Examples of vText."""
         return [cls("Hello World!")]
-
-    from icalendar.param import VALUE
 
     @classmethod
     def from_jcal(cls, jcal_property: list) -> Self:
@@ -362,8 +362,6 @@ class vCalAddress(str):
         /,
         params: dict[str, Any] | None = None,
     ):
-        if params is None:
-            params = {}
         value = to_unicode(value, encoding=encoding)
         self = super().__new__(cls, value)
         self.params = Parameters(params)
@@ -582,8 +580,6 @@ class vFloat(float):
     params: Parameters
 
     def __new__(cls, *args, params: dict[str, Any] | None = None, **kwargs):
-        if params is None:
-            params = {}
         self = super().__new__(cls, *args, **kwargs)
         self.params = Parameters(params)
         return self
@@ -675,8 +671,6 @@ class vInt(int):
     params: Parameters
 
     def __new__(cls, *args, params: dict[str, Any] | None = None, **kwargs):
-        if params is None:
-            params = {}
         self = super().__new__(cls, *args, **kwargs)
         self.params = Parameters(params)
         return self
@@ -813,8 +807,6 @@ class vCategory:
     def __init__(
         self, c_list: list[str] | str, /, params: dict[str, Any] | None = None
     ):
-        if params is None:
-            params = {}
         if not hasattr(c_list, "__iter__") or isinstance(c_list, str):
             c_list = [c_list]
         self.cats: list[vText | str] = [vText(c) for c in c_list]
@@ -1260,8 +1252,6 @@ class vDatetime(TimeBase):
     params: Parameters
 
     def __init__(self, dt, /, params: dict[str, Any] | None = None):
-        if params is None:
-            params = {}
         self.dt = dt
         self.params = Parameters(params)
         self.params.update_tzid_from(dt)
@@ -1429,9 +1419,9 @@ class vDuration(TimeBase):
     default_value: ClassVar[str] = "DURATION"
     params: Parameters
 
-    def __init__(self, td: timedelta, /, params: dict[str, Any] | None = None):
-        if params is None:
-            params = {}
+    def __init__(self, td: timedelta | str, /, params: dict[str, Any] | None = None):
+        if isinstance(td, str):
+            td = vDuration.from_ical(td)
         if not isinstance(td, timedelta):
             raise TypeError("Value MUST be a timedelta instance")
         self.td = td
@@ -1469,7 +1459,7 @@ class vDuration(TimeBase):
     def from_ical(ical):
         match = DURATION_REGEX.match(ical)
         if not match:
-            raise ValueError(f"Invalid iCalendar duration: {ical}")
+            raise InvalidCalendar(f"Invalid iCalendar duration: {ical}")
 
         sign, weeks, days, hours, minutes, seconds = match.groups()
         value = timedelta(
@@ -1786,8 +1776,6 @@ class vWeekday(str):
         /,
         params: dict[str, Any] | None = None,
     ):
-        if params is None:
-            params = {}
         value = to_unicode(value, encoding=encoding)
         self = super().__new__(cls, value)
         match = WEEKDAY_RULE.match(self)
@@ -1857,8 +1845,6 @@ class vFrequency(str):
         /,
         params: dict[str, Any] | None = None,
     ):
-        if params is None:
-            params = {}
         value = to_unicode(value, encoding=encoding)
         self = super().__new__(cls, value)
         if self not in vFrequency.frequencies:
@@ -1923,8 +1909,6 @@ class vMonth(int):
     params: Parameters
 
     def __new__(cls, month: Union[str, int], /, params: dict[str, Any] | None = None):
-        if params is None:
-            params = {}
         if isinstance(month, vMonth):
             return cls(month.to_ical().decode())
         if isinstance(month, str):
@@ -2172,8 +2156,6 @@ class vRecur(CaselessDict):
     )
 
     def __init__(self, *args, params: dict[str, Any] | None = None, **kwargs):
-        if params is None:
-            params = {}
         if args and isinstance(args[0], str):
             # we have a string as an argument.
             args = (self.from_ical(args[0]),) + args[1:]
@@ -2493,19 +2475,21 @@ class vUri(str):
         When a property parameter value is a URI value type, the URI MUST
         be specified as a quoted-string value.
 
-        Example:
-            The following is a URI for a network file:
+    Examples:
+        The following is a URI for a network file:
 
-            .. code-block:: text
+        .. code-block:: text
 
-                http://example.com/my-report.txt
+            http://example.com/my-report.txt
 
-            .. code-block:: pycon
+        .. code-block:: pycon
 
-                >>> from icalendar.prop import vUri
-                >>> uri = vUri.from_ical('http://example.com/my-report.txt')
-                >>> uri
-                'http://example.com/my-report.txt'
+            >>> from icalendar.prop import vUri
+            >>> uri = vUri.from_ical('http://example.com/my-report.txt')
+            >>> uri
+            vUri('http://example.com/my-report.txt')
+            >>> uri.uri
+            'http://example.com/my-report.txt'
     """
 
     default_value: ClassVar[str] = "URI"
@@ -2519,8 +2503,6 @@ class vUri(str):
         /,
         params: dict[str, Any] | None = None,
     ):
-        if params is None:
-            params = {}
         value = to_unicode(value, encoding=encoding)
         self = super().__new__(cls, value)
         self.params = Parameters(params)
@@ -2541,8 +2523,6 @@ class vUri(str):
         """Examples of vUri."""
         return [cls("http://example.com/my-report.txt")]
 
-    from icalendar.param import VALUE
-
     def to_jcal(self, name: str) -> list:
         """The jCal representation of this property according to :rfc:`7265`."""
         return [name, self.params.to_jcal(), self.VALUE.lower(), str(self)]
@@ -2555,6 +2535,107 @@ class vUri(str):
             jcal_property[3],
             Parameters.from_jcal_property(jcal_property),
         )
+
+    @property
+    def ical_value(self) -> str:
+        """The URI."""
+        return self.uri
+
+    @property
+    def uri(self) -> str:
+        """The URI."""
+        return str(self)
+
+    def __repr__(self) -> str:
+        """repr(self)"""
+        return f"{self.__class__.__name__}({self.uri!r})"
+
+    from icalendar.param import FMTTYPE, GAP, LABEL, LANGUAGE, LINKREL, RELTYPE, VALUE
+
+
+class vUid(vText):
+    """A UID of a component.
+
+    This is defined in :rfc:`9253`, Section 7.
+    """
+
+    default_value: ClassVar[str] = "UID"
+
+    @classmethod
+    def new(cls):
+        """Create a new UID for convenience.
+
+        .. code-block:: pycon
+
+            >>> from icalendar import vUid
+            >>> vUid.new()
+            vUid('d755cef5-2311-46ed-a0e1-6733c9e15c63')
+
+        """
+        return vUid(uuid.uuid4())
+
+    @property
+    def uid(self) -> str:
+        """The uid of this property."""
+        return str(self)
+
+    @property
+    def ical_value(self) -> str:
+        """The uid of this property."""
+        return self.uid
+
+    def __repr__(self) -> str:
+        """repr(self)"""
+        return f"{self.__class__.__name__}({self.uid!r})"
+
+    from icalendar.param import FMTTYPE, LABEL, LINKREL
+
+    @classmethod
+    def examples(cls) -> list[vUid]:
+        """Examples of vUid."""
+        return [cls("d755cef5-2311-46ed-a0e1-6733c9e15c63")]
+
+
+class vXmlReference(vUri):
+    """An XML-REFERENCE.
+
+    The associated value references an associated XML artifact and
+    is a URI with an XPointer anchor value.
+
+    This is defined in :rfc:`9253`, Section 7.
+    """
+
+    default_value: ClassVar[str] = "XML-REFERENCE"
+
+    @property
+    def xml_reference(self) -> str:
+        """The XML reference URI of this property."""
+        return self.uri
+
+    @property
+    def x_pointer(self) -> str | None:
+        """The XPointer of the URI.
+
+        The XPointer is defined in `W3C.WD-xptr-xpointer-20021219
+        <https://www.rfc-editor.org/rfc/rfc9253.html#W3C.WD-xptr-xpointer-20021219>`_,
+        and its use as an anchor is defined in `W3C.REC-xptr-framework-20030325
+        <https://www.rfc-editor.org/rfc/rfc9253.html#W3C.REC-xptr-framework-20030325>`_.
+
+        Returns:
+            The decoded x-pointer or ``None`` if not valid x-pointer is found.
+        """
+        from urllib.parse import unquote, urlparse
+
+        parsed = urlparse(self.xml_reference)
+        fragment = unquote(parsed.fragment)
+        if not fragment.startswith("xpointer(") or not fragment.endswith(")"):
+            return None
+        return fragment[9:-1]
+
+    @classmethod
+    def examples(cls) -> list[vXmlReference]:
+        """Examples of vXmlReference."""
+        return [cls("http://example.com/doc.xml#xpointer(/doc/element)")]
 
 
 class vGeo:
@@ -2630,8 +2711,6 @@ class vGeo:
         Raises:
             ValueError: if geo is not a tuple of (latitude, longitude)
         """
-        if params is None:
-            params = {}
         try:
             latitude, longitude = (geo[0], geo[1])
             latitude = float(latitude)
@@ -2751,8 +2830,6 @@ class vUTCOffset:
     # propagate upwards
 
     def __init__(self, td: timedelta, /, params: dict[str, Any] | None = None):
-        if params is None:
-            params = {}
         if not isinstance(td, timedelta):
             raise TypeError("Offset value MUST be a timedelta instance")
         self.td = td
@@ -2872,8 +2949,6 @@ class vInline(str):
         /,
         params: dict[str, Any] | None = None,
     ):
-        if params is None:
-            params = {}
         value = to_unicode(value, encoding=encoding)
         self = super().__new__(cls, value)
         self.params = Parameters(params)
@@ -2945,6 +3020,9 @@ class TypesFactory(CaselessDict):
             vUri,
             vWeekday,
             vCategory,
+            vUid,
+            vXmlReference,
+            vUnknown,
         )
         self["binary"] = vBinary
         self["boolean"] = vBoolean
@@ -2964,7 +3042,9 @@ class TypesFactory(CaselessDict):
         self["inline"] = vInline
         self["date-time-list"] = vDDDLists
         self["categories"] = vCategory
-        self["unknown"] = vUnknown
+        self["unknown"] = vUnknown  # RFC 7265
+        self["uid"] = vUid  # RFC 9253
+        self["xml-reference"] = vXmlReference  # RFC 9253
 
     #################################################
     # Property types
@@ -2992,6 +3072,12 @@ class TypesFactory(CaselessDict):
             "resources": "text",
             "status": "text",
             "summary": "text",
+            # RFC 9253
+            # link should be uri, xml-reference or uid
+            # uri is likely most helpful if people forget to set VALUE
+            "link": "uri",
+            "concept": "uri",
+            "refid": "text",
             # Date and Time Component Properties
             "completed": "date-time",
             "dtend": "date-time",
@@ -3055,6 +3141,10 @@ class TypesFactory(CaselessDict):
             "rsvp": "boolean",
             "sent-by": "cal-address",
             "value": "text",
+            # rfc 9253 parameters
+            "label": "text",
+            "linkrel": "text",
+            "gap": "duration",
         }
     )
 
@@ -3062,9 +3152,9 @@ class TypesFactory(CaselessDict):
         """Returns the type class for a property or parameter.
 
         Args:
-            name: Property or parameter name.
-            value_param: Optional ``VALUE`` parameter, for example, "DATE", "DATE-TIME",
-                or other string.
+            name: Property or parameter name
+            value_param: Optional ``VALUE`` parameter, for example,
+            "DATE", "DATE-TIME", or other string.
 
         Returns:
             The appropriate value type class.
@@ -3121,6 +3211,8 @@ VPROPERTY: TypeAlias = Union[
     vBinary,
     vGeo,
     vUnknown,
+    vXmlReference,
+    vUid,
 ]
 
 __all__ = [
@@ -3152,7 +3244,9 @@ __all__ = [
     "vText",
     "vTime",
     "vUTCOffset",
+    "vUid",
     "vUnknown",
     "vUri",
     "vWeekday",
+    "vXmlReference",
 ]